cfx_rpc_cfx_api

Trait DebugRpcServer

pub trait DebugRpcServer:
    Sized
    + Send
    + Sync
    + 'static {

    // Required methods
    fn txpool_inspect(
        &self,
        address: Option<RpcAddress>,
    ) -> JsonRpcResult<BTreeMap<String, BTreeMap<String, BTreeMap<usize, Vec<String>>>>>;
    fn txpool_content(
        &self,
        address: Option<RpcAddress>,
    ) -> JsonRpcResult<BTreeMap<String, BTreeMap<String, BTreeMap<usize, Vec<RpcTransaction>>>>>;
    fn txpool_get_account_transactions(
        &self,
        address: RpcAddress,
    ) -> JsonRpcResult<Vec<RpcTransaction>>;
    fn txpool_clear(&self) -> JsonRpcResult<()>;
    fn net_throttling(&self) -> JsonRpcResult<Service>;
    fn net_node(&self, node_id: NodeId) -> JsonRpcResult<Option<(String, Node)>>;
    fn net_disconnect_node(
        &self,
        id: NodeId,
        op: Option<UpdateNodeOperation>,
    ) -> JsonRpcResult<bool>;
    fn net_sessions(
        &self,
        node_id: Option<NodeId>,
    ) -> JsonRpcResult<Vec<SessionDetails>>;
    fn current_sync_phase(&self) -> JsonRpcResult<String>;
    fn consensus_graph_state(&self) -> JsonRpcResult<ConsensusGraphStates>;
    fn sync_graph_state(&self) -> JsonRpcResult<SyncGraphStates>;
    fn stat_on_gas_load(
        &self,
        last_epoch: EpochNumber,
        time_window: U64,
    ) -> JsonRpcResult<Option<StatOnGasLoad>>;
    fn transactions_by_epoch(
        &self,
        epoch_number: U64,
    ) -> JsonRpcResult<Vec<WrapTransaction>>;
    fn transactions_by_block(
        &self,
        block_hash: H256,
    ) -> JsonRpcResult<Vec<WrapTransaction>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }

}

Server trait implementation for the DebugRpc RPC API.

Required Methods

fn txpool_inspect( &self, address: Option<RpcAddress>, ) -> JsonRpcResult<BTreeMap<String, BTreeMap<String, BTreeMap<usize, Vec<String>>>>>

fn txpool_content( &self, address: Option<RpcAddress>, ) -> JsonRpcResult<BTreeMap<String, BTreeMap<String, BTreeMap<usize, Vec<RpcTransaction>>>>>

fn txpool_get_account_transactions( &self, address: RpcAddress, ) -> JsonRpcResult<Vec<RpcTransaction>>

fn txpool_clear(&self) -> JsonRpcResult<()>

fn net_throttling(&self) -> JsonRpcResult<Service>

fn net_node(&self, node_id: NodeId) -> JsonRpcResult<Option<(String, Node)>>

fn net_disconnect_node( &self, id: NodeId, op: Option<UpdateNodeOperation>, ) -> JsonRpcResult<bool>

fn net_sessions( &self, node_id: Option<NodeId>, ) -> JsonRpcResult<Vec<SessionDetails>>

fn current_sync_phase(&self) -> JsonRpcResult<String>

fn consensus_graph_state(&self) -> JsonRpcResult<ConsensusGraphStates>

fn sync_graph_state(&self) -> JsonRpcResult<SyncGraphStates>

fn stat_on_gas_load( &self, last_epoch: EpochNumber, time_window: U64, ) -> JsonRpcResult<Option<StatOnGasLoad>>

fn transactions_by_epoch( &self, epoch_number: U64, ) -> JsonRpcResult<Vec<WrapTransaction>>

fn transactions_by_block( &self, block_hash: H256, ) -> JsonRpcResult<Vec<WrapTransaction>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.