cfx_rpc_cfx_api

Trait TxPoolServer

pub trait TxPoolServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn txpool_status(&self) -> JsonRpcResult<TxPoolStatus>;
    fn txpool_next_nonce(&self, address: RpcAddress) -> JsonRpcResult<U256>;
    fn txpool_transaction_by_address_and_nonce(
        &self,
        address: RpcAddress,
        nonce: U256,
    ) -> JsonRpcResult<Option<RpcTransaction>>;
    fn txpool_pending_nonce_range(
        &self,
        address: RpcAddress,
    ) -> JsonRpcResult<TxPoolPendingNonceRange>;
    fn txpool_tx_with_pool_info(
        &self,
        hash: H256,
    ) -> JsonRpcResult<TxWithPoolInfo>;
    fn account_pending_info(
        &self,
        address: RpcAddress,
    ) -> JsonRpcResult<Option<AccountPendingInfo>>;
    fn account_pending_transactions(
        &self,
        address: RpcAddress,
        maybe_start_nonce: Option<U256>,
        maybe_limit: Option<U64>,
    ) -> JsonRpcResult<AccountPendingTransactions>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the TxPool RPC API.

Required Methods

fn txpool_status(&self) -> JsonRpcResult<TxPoolStatus>

fn txpool_next_nonce(&self, address: RpcAddress) -> JsonRpcResult<U256>

fn txpool_transaction_by_address_and_nonce( &self, address: RpcAddress, nonce: U256, ) -> JsonRpcResult<Option<RpcTransaction>>

fn txpool_pending_nonce_range( &self, address: RpcAddress, ) -> JsonRpcResult<TxPoolPendingNonceRange>

fn txpool_tx_with_pool_info(&self, hash: H256) -> JsonRpcResult<TxWithPoolInfo>

fn account_pending_info( &self, address: RpcAddress, ) -> JsonRpcResult<Option<AccountPendingInfo>>

Get transaction pending info by account address

fn account_pending_transactions( &self, address: RpcAddress, maybe_start_nonce: Option<U256>, maybe_limit: Option<U64>, ) -> JsonRpcResult<AccountPendingTransactions>

Get transaction pending info by account address

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.